rename package and remake build file
