+++ cmd/frontend/internal/context/resolvers/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
     visibility = ["//cmd/frontend:__subpackages__"],
     deps = [
         "//cmd/frontend/graphqlbackend",
-        "//internal/codycontext:context",
+        "//internal/codycontext",
         "//internal/database",
         "//internal/gitserver",
         "//internal/trace",
@@ -30,7 +30,7 @@ go_test(
         "//cmd/frontend/graphqlbackend",
         "//internal/actor",
         "//internal/api",
-        "//internal/codycontext:context",
+        "//internal/codycontext",
         "//internal/conf",
         "//internal/database",
         "//internal/database/dbtest",
