+++ internal/codycontext/BUILD.bazel
@@ -1,8 +1,8 @@
-load("//dev:go_defs.bzl", "go_test")
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("//dev:go_defs.bzl", "go_test")
 
 go_library(
-    name = "context",
+    name = "codycontext",
     srcs = [
         "context.go",
         "filter.go",
