+++ internal/codycontext/context.go
@@ -1,4 +1,4 @@
-package context
+package codycontext
 
 import (
 	"context"
